RGB pal testing

+++ far2l/src/console/savescr.cpp
@@ -275,7 +275,7 @@ void SaveScreen::CharCopy(PCHAR_INFO ToBuffer, PCHAR_INFO FromBuffer, int Count)
 
 void SaveScreen::CleanupBuffer(PCHAR_INFO Buffer, size_t BufSize)
 {
-	WORD Attr = FarColorToReal(COL_COMMANDLINEUSERSCREEN);
+	uint64_t Attr = FarColorToReal(COL_COMMANDLINEUSERSCREEN);
 
 	for (size_t i = 0; i < BufSize; i++) {
 		Buffer[i].Attributes = Attr;
